requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16.0,<3.17.0